+++ src/main/java/com/sprint/mission/discodeit/service/basic/BasicMessageService.java
@@ -20,14 +20,16 @@ import com.sprint.mission.discodeit.storage.BinaryContentStorage;
 import java.time.Instant;
 import java.util.List;
 import java.util.NoSuchElementException;
+import java.util.Optional;
 import java.util.UUID;
 import lombok.RequiredArgsConstructor;
+import org.springframework.data.domain.Pageable;
+import org.springframework.data.domain.Slice;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 @RequiredArgsConstructor
 @Service
-@Transactional
 public class BasicMessageService implements MessageService {
 
   private final MessageRepository messageRepository;
@@ -39,7 +41,8 @@ public class BasicMessageService implements MessageService {
   private final PageResponseMapper pageResponseMapper;
 
   @Override
-  public Message create(
+  @Transactional
+  public MessageDto create(
       MessageCreateRequest messageCreateRequest,
       List<BinaryContentCreateRequest> binaryContentCreateRequests
   ) {
@@ -49,7 +52,6 @@ public class BasicMessageService implements MessageService {
     Channel channel = channelRepository.findById(channelId)
         .orElseThrow(() ->
             new NoSuchElementException("Channel with id " + channelId + " does not exist"));
-
     User author = userRepository.findById(authorId)
         .orElseThrow(() ->
             new NoSuchElementException("Author with id " + authorId + " does not exist"));
@@ -60,21 +62,16 @@ public class BasicMessageService implements MessageService {
           String contentType = attachmentRequest.contentType();
           byte[] bytes = attachmentRequest.bytes();
 
-          BinaryContent binaryContent = new BinaryContent(
-              fileName,
-              (long) bytes.length,
-              contentType
-          );
+          BinaryContent binaryContent = new BinaryContent(fileName, (long) bytes.length,
+              contentType);
+          binaryContentRepository.save(binaryContent);
+          binaryContentStorage.put(binaryContent.getId(), bytes);
 
-          BinaryContent savedBinaryContent = binaryContentRepository.save(binaryContent);
-          binaryContentStorage.put(savedBinaryContent.getId(), bytes);
-
-          return savedBinaryContent;
+          return binaryContent;
         })
         .toList();
 
     String content = messageCreateRequest.content();
-
     Message message = new Message(
         content,
         channel,
@@ -82,69 +79,53 @@ public class BasicMessageService implements MessageService {
         attachments
     );
 
-    return messageRepository.save(message);
+    messageRepository.save(message);
+    return messageMapper.toDto(message);
   }
 
   @Override
   @Transactional(readOnly = true)
-  public Message find(UUID messageId) {
+  public MessageDto find(UUID messageId) {
     return messageRepository.findById(messageId)
+        .map(messageMapper::toDto)
         .orElseThrow(() ->
             new NoSuchElementException("Message with id " + messageId + " not found"));
   }
 
   @Override
   @Transactional(readOnly = true)
-  public PageResponse<MessageDto> findAllByChannelId(UUID channelId, Instant cursor) {
-    List<Message> messages;
+  public PageResponse<MessageDto> findAllByChannelId(UUID channelId, Instant createdAt,
+      Pageable pageable) {
+    Slice<MessageDto> slice = messageRepository.findAllByChannelIdWithAuthor(channelId,
+            Optional.ofNullable(createdAt).orElse(Instant.now()),
+            pageable)
+        .map(messageMapper::toDto);
 
-    if (cursor == null) {
-      messages = messageRepository.findTop50ByChannelIdOrderByCreatedAtDesc(channelId);
-    } else {
-      messages = messageRepository
-          .findTop50ByChannelIdAndCreatedAtLessThanOrderByCreatedAtDesc(channelId, cursor);
+    Instant nextCursor = null;
+    if (!slice.getContent().isEmpty()) {
+      nextCursor = slice.getContent().get(slice.getContent().size() - 1)
+          .createdAt();
     }
-
-    List<MessageDto> messageDtos = messages.stream()
-        .map(messageMapper::toDto)
-        .toList();
-
-    Object nextCursor = messages.isEmpty()
-        ? null
-        : messages.get(messages.size() - 1).getCreatedAt();
-
-    boolean hasNext = messages.size() == 50;
-
-    return pageResponseMapper.fromCursor(
-        messageDtos,
-        nextCursor,
-        50,
-        hasNext,
-        null
-    );
+    return pageResponseMapper.fromSlice(slice, nextCursor);
   }
 
   @Override
-  public Message update(UUID messageId, MessageUpdateRequest request) {
+  @Transactional
+  public MessageDto update(UUID messageId, MessageUpdateRequest request) {
     String newContent = request.newContent();
-
     Message message = messageRepository.findById(messageId)
         .orElseThrow(() ->
             new NoSuchElementException("Message with id " + messageId + " not found"));
-
     message.update(newContent);
-
-    return message;
+    return messageMapper.toDto(message);
   }
 
   @Override
+  @Transactional
   public void delete(UUID messageId) {
-    Message message = messageRepository.findById(messageId)
-        .orElseThrow(() ->
-            new NoSuchElementException("Message with id " + messageId + " not found"));
-
-    message.getAttachments()
-        .forEach(binaryContentRepository::delete);
+    if (!messageRepository.existsById(messageId)) {
+      throw new NoSuchElementException("Message with id " + messageId + " not found");
+    }
 
     messageRepository.deleteById(messageId);
   }
